requirements.txt
fastapi
uvicorn
pyTelegramBotAPI
python-dotenv
openai
